Eliminate more inline JS and make a common place for select2 with customized search
