+++ app/javascript/stockaid/index.js
@@ -65,4 +65,3 @@ import "./table_row_click"
 import "./unique_id"
 import "./users"
 import "./utils"
-import "./vendors"
