+++ app/views/purchases/edit.html.erb
@@ -5,10 +5,6 @@
 
   <input type="hidden" id="data-categories" data-value="<%= Category.to_json %>" />
 
-  <%= javascript_tag do %>
-    initializeVendors();
-  <% end %>
-
   <%= form_for @purchase, html: { data: { id: @purchase.id, live_guarded: true } } do |f| %>
     <%= render partial: "purchases/purchase/header", locals: { f: f } %>
     <%= render partial: "purchases/purchase/table", locals: { f: f } %>
