+++ app/views/purchases/new.html.erb
@@ -4,13 +4,7 @@
   <%= render partial: "purchases/templates/purchase_item_options" %>
 
   <input type="hidden" id="data-categories" data-value="<%= Category.to_json %>" />
-
-  <%= javascript_tag do %>
-    var data = data || {};
-    data.vendors = <%= @vendors.to_json.html_safe %>;
-
-    initializeVendors();
-  <% end %>
+  <input type="hidden" id="data-vendors" data-value="<%= @vendors.to_json %>" />
 
   <%= form_for @purchase, html: { data: { id: @purchase.id, live_guarded: true } } do |f| %>
     <%= render partial: "purchases/purchase/header", locals: { f: f } %>
